add DocumentMapper and update entites

+++ src/main/java/com/paperless/persistence/repositories/AuthtokenTokenRepository.java
@@ -1,9 +1,9 @@
 package com.paperless.persistence.repositories;
 
-import at.fhtw.swkom.paperless.persistance.entities.AuthtokenTokenEntity;
+import com.paperless.persistence.entities.AuthtokenTokenEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuthtokenTokenRepository extends JpaRepository<AuthtokenTokenEntity,Integer> {
+public interface AuthtokenTokenRepository extends JpaRepository<AuthtokenTokenEntity,Long> {
 }
